+++ src/wagtail_vector_index/ai_utils/types.py
